# api.yml
openapi: 3.0.1
info:
  title: ExtranetAPI
  version: '1.0'
paths:
  /analytics/margin/{projectId}:
    get:
      tags:
        - Analytics
      summary: Маржа по проекту
      parameters:
        - name: projectId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Маржа по проекту
          content:
            text/plain:
              schema:
                type: number
                format: double
            application/json:
              schema:
                type: number
                format: double
            text/json:
              schema:
                type: number
                format: double
  /analytics/margin:
    post:
      tags:
        - Analytics
      summary: Маржа по проектам за период
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchAnalyticDto'
          text/json:
            schema:
              $ref: '#/components/schemas/SearchAnalyticDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/SearchAnalyticDto'
      responses:
        '200':
          description: Маржа по проектам за период
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProjectsMagrinDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectsMagrinDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ProjectsMagrinDto'
  /analytics/profitNorm/{projectId}:
    get:
      tags:
        - Analytics
      summary: Норма прибыли
      parameters:
        - name: projectId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Норма прибыли
          content:
            text/plain:
              schema:
                type: string
            application/json:
              schema:
                type: string
            text/json:
              schema:
                type: string
  /analytics/rateOfSurplusValue/{projectId}:
    get:
      tags:
        - Analytics
      summary: Норма прибавочной стоимости
      parameters:
        - name: projectId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Норма прибавочной стоимости
          content:
            text/plain:
              schema:
                type: string
            application/json:
              schema:
                type: string
            text/json:
              schema:
                type: string
  /analytics/outDatedProjects:
    post:
      tags:
        - Analytics
      summary: Просроченные проекты
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchAnalyticDto'
          text/json:
            schema:
              $ref: '#/components/schemas/SearchAnalyticDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/SearchAnalyticDto'
      responses:
        '200':
          description: Просроченные проекты
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/OutdatedProjectsDto'
            application/json:
              schema:
                $ref: '#/components/schemas/OutdatedProjectsDto'
            text/json:
              schema:
                $ref: '#/components/schemas/OutdatedProjectsDto'
  /analytics/spendingOnCosts/{projectId}:
    get:
      tags:
        - Analytics
      summary: Траты на издержки по проекту
      parameters:
        - name: projectId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Траты на издержки по проекту
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/SpendingOnProjectCostsDto'
            application/json:
              schema:
                $ref: '#/components/schemas/SpendingOnProjectCostsDto'
            text/json:
              schema:
                $ref: '#/components/schemas/SpendingOnProjectCostsDto'
  /analytics/spendingOnCosts:
    post:
      tags:
        - Analytics
      summary: Траты на издержки
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchAnalyticDto'
          text/json:
            schema:
              $ref: '#/components/schemas/SearchAnalyticDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/SearchAnalyticDto'
      responses:
        '200':
          description: Траты на издержки
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SpendingOnCostDto'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SpendingOnCostDto'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SpendingOnCostDto'
  /analytics/projectsPrices:
    post:
      tags:
        - Analytics
      summary: Цены проектов
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Period'
          text/json:
            schema:
              $ref: '#/components/schemas/Period'
          application/*+json:
            schema:
              $ref: '#/components/schemas/Period'
      responses:
        '200':
          description: Цены проектов
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProjectsPricesDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectsPricesDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ProjectsPricesDto'
...
components:
  schemas:
...
    OutdatedProjectDto:
      type: object
      properties:
        startDate:
          type: string
          format: date-time
        deadLine:
          type: string
          format: date-time
        projectName:
          type: string
          nullable: true
        wastedDays:
          type: integer
          format: int32
      additionalProperties: false
    OutdatedProjectsDto:
      type: object
      properties:
        outdatedProjects:
          type: array
          items:
            $ref: '#/components/schemas/OutdatedProjectDto'
          nullable: true
        period:
          $ref: '#/components/schemas/Period'
        averageAmount:
          type: number
          format: double
      additionalProperties: false
    Period:
      type: object
      properties:
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
      additionalProperties: false
...
    ProjectMarginDto:
      type: object
      properties:
        projectStartdate:
          type: string
          format: date-time
        projectDeadLine:
          type: string
          format: date-time
        projectName:
          type: string
          nullable: true
        projectMargin:
          type: number
          format: double
      additionalProperties: false
    ProjectPriceDto:
      type: object
      properties:
        projectName:
          type: string
          nullable: true
        projectPrice:
          type: number
          format: double
      additionalProperties: false
    ProjectsMagrinDto:
      type: object
      properties:
        projectMargins:
          type: array
          items:
            $ref: '#/components/schemas/ProjectMarginDto'
          nullable: true
        period:
          $ref: '#/components/schemas/Period'
        totalMargin:
          type: number
          format: double
      additionalProperties: false
    ProjectsPricesDto:
      type: object
      properties:
        averagePrice:
          type: number
          format: double
        projectPrices:
          type: array
          items:
            $ref: '#/components/schemas/ProjectPriceDto'
          nullable: true
        maxProjectPrice:
          $ref: '#/components/schemas/ProjectPriceDto'
        minProjectPrice:
          $ref: '#/components/schemas/ProjectPriceDto'
      additionalProperties: false
...
    SearchAnalyticDto:
      type: object
      properties:
        name:
          type: string
          nullable: true
        period:
          $ref: '#/components/schemas/Period'
      additionalProperties: false
    SpendingOnCostDto:
      type: object
      properties:
        name:
          type: string
          nullable: true
        amount:
          type: number
          format: double
      additionalProperties: false
    SpendingOnProjectCostsDto:
      type: object
      properties:
        spendingOnCosts:
          type: array
          items:
            $ref: '#/components/schemas/SpendingOnCostDto'
          nullable: true
        projectName:
          type: string
          nullable: true
      additionalProperties: false
...